Refactor scope data into a module per service
Rather than defining `scopes.data` as a single module listing all
services, break it apart per-service. This makes maintenance easier
for more complex services like GCS.
