+++ src/globus_sdk/scopes/data/__init__.py
@@ -1,0 +1,19 @@
+from .auth import AuthScopes
+from .flows import FlowsScopes
+from .gcs import GCSCollectionScopeBuilder, GCSEndpointScopeBuilder
+from .groups import GroupsScopes, NexusScopes
+from .search import SearchScopes
+from .timer import TimerScopes
+from .transfer import TransferScopes
+
+__all__ = (
+    "AuthScopes",
+    "FlowsScopes",
+    "GCSEndpointScopeBuilder",
+    "GCSCollectionScopeBuilder",
+    "GroupsScopes",
+    "NexusScopes",
+    "SearchScopes",
+    "TimerScopes",
+    "TransferScopes",
+)
