+++ src/globus_sdk/scopes/data/search.py
@@ -1,0 +1,15 @@
+from ..builder import ScopeBuilder
+
+SearchScopes = ScopeBuilder(
+    "search.api.globus.org",
+    known_scopes=[
+        "all",
+        "globus_connect_server",
+        "ingest",
+        "search",
+    ],
+)
+"""Globus Search scopes.
+
+.. listknownscopes:: globus_sdk.scopes.SearchScopes
+"""
